backend-python/requirements.txt
Flask
flask
google-cloud-datastore==2.8.2